frontend/requirements.txt
django
hrflow
python-dotenv